add example of learning rate scheduler

+++ vision_utils/custom_torch_utils.py
@@ -11,7 +11,7 @@ from tensorboardX import SummaryWriter
 import os
 import glob
 import shutil
-
+from ignite.contrib.handlers.param_scheduler import CosineAnnealingScheduler
 
 def set_parameter_requires_grad(model, feature_extracting):
     if feature_extracting:
@@ -118,7 +118,8 @@ def create_summary_writer(model, data_loader, log_dir):
 def run(path_to_model_script, epochs, log_interval, dataloaders,
         dirname='resnet_models', filename_prefix='resnet', n_saved=2,
         log_dir='../../fer2013/logs', launch_tensorboard=False, patience=10,
-        resume_model=None, resume_optimizer=None, backup_step=1, backup_path=''):
+        resume_model=None, resume_optimizer=None, backup_step=1, backup_path=None,
+        lr_start=None, lr_end=None):
 
     if launch_tensorboard:
         os.makedirs(log_dir, exist_ok=True)
@@ -135,6 +136,7 @@ def run(path_to_model_script, epochs, log_interval, dataloaders,
 
     model = model_script['my_model']
     optimizer = model_script['optimizer']
+    lr_schedulers = model_script.get('lr_schedulers', None)
 
     if resume_model:
         model.load_state_dict(torch.load(resume_model))
@@ -154,20 +156,14 @@ def run(path_to_model_script, epochs, log_interval, dataloaders,
                                                      'nll': Loss(F.cross_entropy)},
                                             device=device)
 
-    def get_val_loss(engine):
-        return -engine.state.metrics['nll']
-
-    checkpointer = handlers.ModelCheckpoint(dirname=dirname, filename_prefix=filename_prefix,
-                                            score_function=get_val_loss,
-                                            score_name='val_loss',
-                                            n_saved=n_saved, create_dir=True,
-                                            require_empty=False, save_as_state_dict=True
-                                            )
-    earlystop = handlers.EarlyStopping(patience, get_val_loss, trainer)
-
-    evaluator.add_event_handler(Events.EPOCH_COMPLETED, checkpointer,
-                                {'optimizer': optimizer, 'model': model})
-    evaluator.add_event_handler(Events.EPOCH_COMPLETED, earlystop)
+    if lr_schedulers is not None:
+        for sched in lr_schedulers:
+            sched.cycle_size = len(train_loader)
+            if lr_start is not None:
+                sched.start_value = lr_start
+            if lr_end is not None:
+                sched.end_value = lr_end
+            trainer.add_event_handler(Events.ITERATION_STARTED, sched)
 
     desc = "ITERATION - loss: {:.3f}"
     pbar = tqdm(
@@ -217,27 +213,45 @@ def run(path_to_model_script, epochs, log_interval, dataloaders,
         if launch_tensorboard:
             val_writer.add_scalar('avg_loss', avg_nll, engine.state.epoch)
             val_writer.add_scalar('avg_accuracy', avg_accuracy, engine.state.epoch)
+        return -avg_nll
+
+    # def get_val_loss(engine):
+    #     return -engine.state.metrics['nll']
+
+    checkpointer = handlers.ModelCheckpoint(dirname=dirname, filename_prefix=filename_prefix,
+                                            # score_function=get_val_loss,
+                                            score_function=log_validation_results,
+                                            score_name='val_loss',
+                                            n_saved=n_saved, create_dir=True,
+                                            require_empty=False, save_as_state_dict=True
+                                            )
+    earlystop = handlers.EarlyStopping(patience, log_validation_results, trainer)
+
+    evaluator.add_event_handler(Events.EPOCH_COMPLETED, checkpointer,
+                                {'optimizer': optimizer, 'model': model})
+    evaluator.add_event_handler(Events.EPOCH_COMPLETED, earlystop)
 
     # optimizer and model that are in the gdrive, created from a previous run
     original_files = glob.glob(os.path.join(backup_path, '*.pth*'))
 
     @trainer.on(Events.EPOCH_COMPLETED)
     def backup_checkpoints(engine):
-        if engine.state.epoch % backup_step == 0:
+        if backup_path is not None:
+            if engine.state.epoch % backup_step == 0:
 
-            # get old model and optimizer files paths so that we can remove them after copying the newer ones
-            old_files = glob.glob(os.path.join(backup_path, '*.pth'))
+                # get old model and optimizer files paths so that we can remove them after copying the newer ones
+                old_files = glob.glob(os.path.join(backup_path, '*.pth'))
 
-            # get new model and optimizer checkpoints
-            new_files = glob.glob(os.path.join(dirname, '*.pth*'))
-            if len(new_files) > 0:  # copy new checkpoints from local checkpoint folder to the backup_path folder
-                for f_ in new_files:
-                    shutil.copy2(f_, backup_path)
+                # get new model and optimizer checkpoints
+                new_files = glob.glob(os.path.join(dirname, '*.pth*'))
+                if len(new_files) > 0:  # copy new checkpoints from local checkpoint folder to the backup_path folder
+                    for f_ in new_files:
+                        shutil.copy2(f_, backup_path)
 
-                if len(old_files) > 0:  # remove older checkpoints as the new ones have been copied
-                    for f_ in old_files:
-                        if f_ not in original_files:
-                            os.remove(f_)
+                    if len(old_files) > 0:  # remove older checkpoints as the new ones have been copied
+                        for f_ in old_files:
+                            if f_ not in original_files:
+                                os.remove(f_)
 
     if launch_tensorboard:
         @trainer.on(Events.EPOCH_COMPLETED)
